tests rewrite / free from prefix
        - tests rewritten with plone.testing layers
        - not tied anymore to sys.prefix

+++ src/minitage/core/tests/unpackers/test_tar.py
@@ -1,32 +1,3 @@
-# Copyright (C) 2009, Mathieu PASQUET <kiorky@cryptelium.net>
-# All rights reserved.
-#
-# Redistribution and use in source and binary forms, with or without
-# modification, are permitted provided that the following conditions are met:
-#
-# 1. Redistributions of source code must retain the above copyright notice,
-#    this list of conditions and the following disclaimer.
-# 2. Redistributions in binary form must reproduce the above copyright
-#    notice, this list of conditions and the following disclaimer in the
-#    documentation and/or other materials provided with the distribution.
-# 3. Neither the name of the <ORGANIZATION> nor the names of its
-#    contributors may be used to endorse or promote products derived from
-#    this software without specific prior written permission.
-#
-# THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
-# AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-# IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
-# ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE
-# LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR
-# CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF
-# SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS
-# INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN
-# CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
-# ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
-# POSSIBILITY OF SUCH DAMAGE.
-
-
-
 __docformat__ = 'restructuredtext en'
 
 import unittest
@@ -36,26 +7,30 @@ import tempfile
 
 from minitage.core.unpackers import interfaces
 
-opts = dict(
-    path=os.path.expanduser('~/minitagerepo'),
-    dest=os.path.expanduser('~/minitagerepodest'),
-    wc=os.path.expanduser('~/minitagerepodestwc'),
-)
-
-prefix = os.getcwd()
-
+opts = dict()
 path = tempfile.mkdtemp()
 
-class testTar(unittest.TestCase):
+from minitage.core.tests import base
+
+class testTar(base.TestCase):
     """testtar"""
 
-    def tearDown(self):
-        """."""
-        shutil.rmtree(path)
-        os.makedirs(path)
+    def setUp(self):
+        opts.update(dict(
+            path=os.path.expanduser('%(p)s/minitagerepo') % self.layer,
+            dest=os.path.expanduser('%(p)s/minitagerepodest') % self.layer,
+            wc=os.path.expanduser('%(p)s/minitagerepodestwc') % self.layer,
+        ))
+        for k in opts:
+            path = opts[k]
+            if os.path.exists(path):
+                shutil.rmtree(path)
+            os.makedirs(path)
+
 
     def testTarfile(self):
         """testTarfile."""
+        path = opts['path']
         os.chdir(path)
         os.system("""
                   mkdir a b;
@@ -73,6 +48,7 @@ class testTar(unittest.TestCase):
 
     def testTarbz2file(self):
         """testTarbz2file."""
+        path = opts['path']
         os.chdir(path)
         os.system("""
                   mkdir a;
@@ -88,6 +64,7 @@ class testTar(unittest.TestCase):
 
     def testTargzfile(self):
         """testTargzfile."""
+        path = opts['path']
         os.chdir(path)
         os.system("""
                   mkdir a;
@@ -101,9 +78,10 @@ class testTar(unittest.TestCase):
         self.assertTrue(os.path.isfile('a/toto'))
         self.assertEquals(open('a/toto').read(),'aaaa\n')
 
-if __name__ == '__main__':
+
+def test_suite():
     suite = unittest.TestSuite()
     suite.addTest(unittest.makeSuite(testTar))
-    unittest.TextTestRunner(verbosity=2).run(suite)
+    return suite
 
 # vim:set et sts=4 ts=4 tw=80:
